# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Flash-Archi Public API
  description: >
    API publique de Flash-Archi (pré-production) — projets, jobs,
    artefacts, versions et webhooks.

    Clés d'accès :
    - **API publique** (`/api/public/v1/**`) : clé Bearer étendue avec scopes
      `public:*`, résolue par tenant (plan : free / pro / enterprise).
    - **API interne** (`/api/flash-archi/projects/**/versions/**`)
      : clé applicative `API_KEY` (environnement), scopes internes.

    Tous les endpoints honorent le rate limiting (headers `X-RateLimit-*`).
  version: 3.3.0
  contact:
    name: Flash-Archi
servers:
- url: /
  description: "Racine \u2014 chemins complets (API publique + interne)"
...
paths:
  /api/public/v1/projects:
    get:
      tags:
      - Projets
      summary: Lister les projets du tenant
      operationId: listProjects
      security:
      - PublicKeyAuth:
        - public:read:projects
      parameters:
      - $ref: '#/components/parameters/PaginationLimit'
      - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: "Liste des projets (isol\xE9e par tenant via RLS)"
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - limit
                - offset
                - count
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  limit:
                    type: integer
                  offset:
                    type: integer
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      tags:
      - Projets
      summary: "Cr\xE9er un projet"
      operationId: createProject
      security:
      - PublicKeyAuth:
        - public:write:projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: "Projet cr\xE9\xE9"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/public/v1/webhooks:
    get:
      tags:
      - Webhooks
      summary: Lister les abonnements webhook (sans secrets)
      operationId: listWebhooks
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      responses:
        '200':
          description: Liste des webhooks du tenant
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookSubscription'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      tags:
      - Webhooks
      summary: "Créer un abonnement webhook (le secret est retourné une seule fois)"
      operationId: createWebhook
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - url
              - events
              properties:
                url:
                  type: string
                  format: uri
                  description: "URL de r\xE9ception du webhook"
                  example: https://hooks.example.com/fa
                events:
                  type: array
                  items:
                    type: string
                    enum:
                    - job.completed
                    - job.failed
                    - project.created
                    - project.updated
                  minItems: 1
                  example:
                  - job.completed
                  - job.failed
      responses:
        '201':
          description: "Webhook cr\xE9\xE9"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookWithSecret'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/flash-archi/projects/{projectId}/versions:
    get:
      tags:
      - Versions
      summary: Lister les versions d'un projet (API interne)
      operationId: listVersions
      security:
      - InternalKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/PaginationLimit'
      - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Liste des versions
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - count
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Version'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - Versions
      summary: "Cr\xE9er une version (API interne)"
      operationId: createVersion
      security:
      - InternalKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionInput'
      responses:
        '201':
          description: "Version cr\xE9\xE9e"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/flash-archi/projects/{projectId}/versions/{versionId}:
    get:
      tags:
      - Versions
      summary: Lire une version (API interne)
      operationId: getVersion
      security:
      - InternalKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/flash-archi/projects/{projectId}/versions/{versionId}/restore:
    post:
      tags:
      - Versions
      summary: Restaurer une version (API interne)
      operationId: restoreVersion
      security:
      - InternalKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - message
              properties:
                message:
                  type: string
                  description: "Message associ\xE9 \xE0 la restauration"
                  example: "Retour \xE0 la version 2"
      responses:
        '200':
          description: "Restauration r\xE9ussie"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/flash-archi/projects/{projectId}/versions/{versionA}/compare/{versionB}:
    get:
      tags:
      - Versions
      summary: Comparer deux versions (API interne)
      operationId: compareVersions
      security:
      - InternalKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: versionA
        in: path
        required: true
        schema:
          type: string
        description: "ID de la premi\xE8re version"
      - name: versionB
        in: path
        required: true
        schema:
          type: string
        description: ID de la seconde version
      responses:
        '200':
          description: "Diff\xE9rence entre les deux versions"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    PublicKeyAuth:
      type: http
      scheme: bearer
      description: "Cl\xE9 API publique (scope `public:*`) r\xE9solue par tenant.
        \ Le scope requis est indiqu\xE9 par endpoint. Plans : free=100 req/min, pro=500
        \ req/min, enterprise=1000 req/min.\n"
    InternalKeyAuth:
      type: http
      scheme: bearer
      description: "Cl\xE9 applicative interne `API_KEY` (routes versions)."
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Identifiant du projet
...
    VersionId:
      name: versionId
      in: path
      required: true
      schema:
        type: string
      description: Identifiant de la version
...
    PaginationLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: "Nombre d'\xE9l\xE9ments (max 100)"
    PaginationOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: "D\xE9calage de pagination"
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prompt:
          type: string
        parameters:
          type: object
          additionalProperties: true
        tenant_id:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required:
      - name
      - prompt
      properties:
        name:
          type: string
          example: Villa contemporaine
        prompt:
          type: string
          example: "Maison familiale de 120 m\xB2 avec garage"
        parameters:
          type: object
          additionalProperties: true
          example:
            levels: 1
            style: contemporain
...
    Version:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
          format: uuid
        version_number:
          type: integer
        state:
          type: object
          additionalProperties: true
        message:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
    VersionInput:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          example: Version initiale
    WebhookSubscription:
      type: object
      description: "Abonnement webhook (le secret n'est JAMAIS expos\xE9)"
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
            - job.completed
            - job.failed
            - project.created
            - project.updated
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WebhookWithSecret:
      allOf:
      - $ref: '#/components/schemas/WebhookSubscription'
      - type: object
        description: "Retourn\xE9 UNIQUEMENT \xE0 la cr\xE9ation"
        required:
        - secret
        properties:
          secret:
            type: string
            description: "Cl\xE9 HMAC-SHA256 \u2014 \xE0 conserver, non r\xE9cup\xE9
              rable ensuite"
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: "Requ\xEAte invalide"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: "Authentification \xE9chou\xE9e ou cl\xE9 manquante"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Scope insuffisant pour cette ressource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: "Ressource introuvable (ou hors du tenant \u2014 RLS, 404 volontaire)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: "Conflit (ex. rollback bloqu\xE9 par une collaboration active)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: "Quota d\xE9pass\xE9"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: "Quota de la fen\xEAtre"
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: "Requ\xEAtes restantes dans la fen\xEAtre"
        X-RateLimit-Reset:
          schema:
            type: integer
          description: "Timestamp (ms) de r\xE9initialisation"
        Retry-After:
          schema:
            type: integer
          description: "Secondes avant retry autoris\xE9"